Update: Move pattern actions.
