+++ packages/edit-site/src/style.scss
@@ -13,7 +13,6 @@
 @import "./components/page-templates/style.scss";
 @import "./components/table/style.scss";
 @import "./components/editor/style.scss";
-@import "./components/create-template-part-modal/style.scss";
 @import "./components/welcome-guide/style.scss";
 @import "./components/layout/style.scss";
 @import "./components/save-hub/style.scss";
