+++ packages/editor/src/components/create-template-part-modal/style.scss
@@ -1,5 +1,5 @@
-.edit-site-create-template-part-modal {
-	z-index: z-index(".edit-site-create-template-part-modal");
+.editor-create-template-part-modal {
+	z-index: z-index(".editor-create-template-part-modal");
 
 	.components-modal__frame {
 		@include break-small {
@@ -8,12 +8,12 @@
 	}
 }
 
-.edit-site-create-template-part-modal__area-radio-group {
+.editor-create-template-part-modal__area-radio-group {
 	width: 100%;
 	border: $border-width solid $gray-700;
 	border-radius: 2px;
 
-	.components-button.edit-site-create-template-part-modal__area-radio {
+	.components-button.editor-create-template-part-modal__area-radio {
 		display: block;
 		width: 100%;
 		height: 100%;
@@ -46,12 +46,12 @@
 			color: $gray-900;
 			cursor: auto;
 
-			.edit-site-create-template-part-modal__option-label div {
+			.editor-create-template-part-modal__option-label div {
 				color: $gray-600;
 			}
 		}
 
-		.edit-site-create-template-part-modal__option-label {
+		.editor-create-template-part-modal__option-label {
 			padding-top: $grid-unit-05;
 			white-space: normal;
 
@@ -61,7 +61,7 @@
 			}
 		}
 
-		.edit-site-create-template-part-modal__checkbox {
+		.editor-create-template-part-modal__checkbox {
 			margin-left: auto;
 			min-width: $grid-unit-30;
 		}
