+++ packages/editor/src/components/post-actions/actions.js
@@ -32,9 +32,11 @@ import { store as editorStore } from '../../store';
 import { unlock } from '../../lock-unlock';
 import isTemplateRevertable from '../../store/utils/is-template-revertable';
 import { exportPatternAsJSONAction } from './export-pattern-action';
+import { CreateTemplatePartModalContents } from '../create-template-part-modal';
 
 // Patterns.
-const { PATTERN_TYPES } = unlock( patternsPrivateApis );
+const { PATTERN_TYPES, CreatePatternModalContents, useDuplicatePatternProps } =
+	unlock( patternsPrivateApis );
 
 /**
  * Check if a template is removable.
@@ -968,6 +970,64 @@ const resetTemplateAction = {
 	},
 };
 
+export const duplicatePatternAction = {
+	id: 'duplicate-pattern',
+	label: _x( 'Duplicate', 'action label' ),
+	isEligible: ( item ) => item.type !== TEMPLATE_PART_POST_TYPE,
+	modalHeader: _x( 'Duplicate pattern', 'action label' ),
+	RenderModal: ( { items, closeModal } ) => {
+		const [ item ] = items;
+		const isThemePattern = item.type === PATTERN_TYPES.theme;
+		const duplicatedProps = useDuplicatePatternProps( {
+			pattern: isThemePattern ? item : item.patternPost,
+			onSuccess: () => closeModal(),
+		} );
+		return (
+			<CreatePatternModalContents
+				onClose={ closeModal }
+				confirmLabel={ _x( 'Duplicate', 'action label' ) }
+				{ ...duplicatedProps }
+			/>
+		);
+	},
+};
+
+export const duplicateTemplatePartAction = {
+	id: 'duplicate-template-part',
+	label: _x( 'Duplicate', 'action label' ),
+	isEligible: ( item ) => item.type === TEMPLATE_PART_POST_TYPE,
+	modalHeader: _x( 'Duplicate template part', 'action label' ),
+	RenderModal: ( { items, closeModal } ) => {
+		const [ item ] = items;
+		const { createSuccessNotice } = useDispatch( noticesStore );
+		function onTemplatePartSuccess() {
+			createSuccessNotice(
+				sprintf(
+					// translators: %s: The new template part's title e.g. 'Call to action (copy)'.
+					__( '"%s" duplicated.' ),
+					item.title
+				),
+				{ type: 'snackbar', id: 'edit-site-patterns-success' }
+			);
+			closeModal();
+		}
+		return (
+			<CreateTemplatePartModalContents
+				blocks={ item.blocks }
+				defaultArea={ item.templatePart.area }
+				defaultTitle={ sprintf(
+					/* translators: %s: Existing template part title */
+					__( '%s (Copy)' ),
+					item.title
+				) }
+				onCreate={ onTemplatePartSuccess }
+				onError={ closeModal }
+				confirmLabel={ _x( 'Duplicate', 'action label' ) }
+			/>
+		);
+	},
+};
+
 export function usePostActions( postType, onActionPerformed ) {
 	const { postTypeObject } = useSelect(
 		( select ) => {
@@ -1001,6 +1061,8 @@ export function usePostActions( postType, onActionPerformed ) {
 				  ! isPattern &&
 				  duplicatePostAction
 				: false,
+			isTemplateOrTemplatePart && duplicateTemplatePartAction,
+			isPattern && duplicatePatternAction,
 			renamePostAction,
 			isPattern && exportPatternAsJSONAction,
 			isTemplateOrTemplatePart ? resetTemplateAction : restorePostAction,
