+++ packages/editor/src/style.scss
@@ -3,6 +3,7 @@
 @import "./components/autocompleters/style.scss";
 @import "./components/block-manager/style.scss";
 @import "./components/collapsible-block-toolbar/style.scss";
+@import "./components/create-template-part-modal/style.scss";
 @import "./components/block-settings-menu/style.scss";
 @import "./components/blog-title/style.scss";
 @import "./components/document-bar/style.scss";
